Split original clipboard functions out
